Change the data type of imdb_score from SMALLINT to REAL to better handle floating values

+++ db/schema/movies.sql
@@ -6,7 +6,7 @@ CREATE TABLE movies (
   release_date DATE NOT NULL,
   rating VARCHAR(255) NOT NULL DEFAULT '0',
   genre VARCHAR(255) NOT NULL,
-  imdb_score SMALLINT NOT NULL DEFAULT '0',
+  imdb_score REAL NOT NULL DEFAULT '0',
   user_id INT REFERENCES users(id) ON DELETE CASCADE
 );
